requirements.txt
openai==1.35.10
python-dotenv==1.0.1
pydub==0.25.1
